clean up websocket code - api is now specific to revision of europi
